requirements.txt
pandas
xgboost
joblib
kafka-python
psycopg2-binary